requirements.txt
Flask==2.3.3
Flask-Mail==0.9.1
python-dotenv==1.0.0
email-validator==2.0.0
Werkzeug==2.3.7
Jinja2==3.1.2